+++ GCD/extended_gcd.h
@@ -3,6 +3,6 @@
 
 
 
-int extended_gcd(int a,int b,int &x,int &y,const int &start=1);
+long long extended_gcd(long long a,long long b,long long &x,long long &y,const long long &start=1);
 #endif
