+++ GCD/line_mod_equotion.h
@@ -2,5 +2,5 @@
 #define _LINE_MOD_EQUTION_H
 #include<vector>
 
-std::vector<int> line_mod_equotion(const int a,const int b,const int n);
+std::vector<long long> line_mod_equotion(const long long a,const long long b,const long long n);
 #endif
